+++ src/main/java/org/ojalgo/optimisation/integer/IntegerStrategy.java
@@ -185,8 +185,10 @@ public interface IntegerStrategy {
         Comparator<NodeKey>[] definitions = (Comparator<NodeKey>[]) new Comparator<?>[] { NodeKey.EARLIEST_SEQUENCE, NodeKey.LARGEST_DISPLACEMENT,
                 NodeKey.SMALLEST_DISPLACEMENT, NodeKey.LATEST_SEQUENCE };
 
-        return new ConfigurableStrategy(Parallelism.CORES.require(4), definitions, NumberContext.of(12, 8), NumberContext.of(7, 8), DefaultStrategy::new,
-                new GMICutConfiguration());
+        NumberContext integrality = NumberContext.of(12, 8);
+        NumberContext gap = NumberContext.of(7, 8);
+
+        return new ConfigurableStrategy(Parallelism.CORES.require(4), definitions, integrality, gap, DefaultStrategy::new, new GMICutConfiguration());
     }
 
     /**